Rename existing cases to expansion_cases
