+++ tests/query/expansion_cases.py
